+++ assets/test-utils/app-context-providers.tsx
@@ -37,6 +37,7 @@ export const DEFAULT_SITE: PlausibleSite = {
   explorationMaxJourneySteps: 0,
   propsAvailable: false,
   siteSegmentsAvailable: false,
+  siteAnnotationsAvailable: false,
   conversionsOptedOut: false,
   funnelsOptedOut: false,
   propsOptedOut: false,