+++ e2e/tests/test-utils.ts
@@ -43,10 +43,15 @@ export const expectHeaders = async (report: Locator, headers: HaveTextArg) =>
   expect(report.getByTestId('report-header')).toHaveText(headers)
 
 export const expectRows = async (report: Locator, labels: HaveTextArg) =>
-  expect(report.getByTestId('report-row').getByRole('link')).toHaveText(labels)
+  expect(
+    report.getByTestId('report-row').getByTestId('dimension-value')
+  ).toHaveText(labels)
 
 export const rowLink = (report: Locator, label: HasTextArg) =>
-  report.getByTestId('report-row').filter({ hasText: label }).getByRole('link')
+  report
+    .getByTestId('report-row')
+    .filter({ hasText: label })
+    .getByTestId('dimension-value')
 
 export const expectMetricValues = async (
   report: Locator,